refactored to LPA helper with tests
