+++ service-front/module/Application/src/Controller/IndexController.php
@@ -6,7 +6,9 @@ namespace Application\Controller;
 
 use Application\Contracts\OpgApiServiceInterface;
 use Application\Exceptions\HttpException;
+use Application\Exceptions\OpgFrontException;
 use Application\Helpers\AddressProcessorHelper;
+use Application\Helpers\LpaHelper;
 use Application\Services\SiriusApiService;
 use DateTime;
 use Laminas\Http\Response;
@@ -25,8 +27,10 @@ class IndexController extends AbstractActionController
 
     public function __construct(
         private readonly OpgApiServiceInterface $opgApiService,
-        private readonly SiriusApiService $siriusApiService
-    ) {
+        private readonly SiriusApiService       $siriusApiService,
+        private readonly LpaHelper              $lpaHelper
+    )
+    {
     }
 
     public function indexAction()
@@ -46,6 +50,11 @@ class IndexController extends AbstractActionController
 
         /** @var string $type */
         $type = $this->params()->fromQuery("personType");
+
+        if (!$this->lpaHelper->lpaIdentitiesMatch($lpas, $type)) {
+            $personTypeDescription = $type === 'donor' ? "Donors" : " Certificate Providers";
+            throw new OpgFrontException("These LPAs are for different " . $personTypeDescription);
+        }
 
         /**
          * @psalm-suppress PossiblyUndefinedArrayOffset
@@ -73,7 +82,7 @@ class IndexController extends AbstractActionController
     private function processLpaResponse(string $type, array $data): array
     {
         if ($type === 'donor') {
-            if (! empty($data['opg.poas.lpastore'])) {
+            if (!empty($data['opg.poas.lpastore'])) {
                 $address = (new AddressProcessorHelper())->processAddress(
                     $data['opg.poas.lpastore']['donor']['address'],
                     'lpaStoreAddressType'
