+++ service-front/module/Application/src/Exceptions/OpgFrontException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Application\Exceptions;
+
+class OpgFrontException extends \Exception
+{
+}
